Change the type of the id field from integer to long

+++ dawn-data-access/src/main/java/cn/uuxl/dawn/dal/mapper/UserMapper.java
@@ -8,6 +8,6 @@ import java.util.List;
 public interface UserMapper {
     
     List<User> selectAll();
-    
-    User selectById(Integer id);
+
+    User selectById(Long id);
 }
